WIP making chip material
