+++ app/src/main/res/drawable-v21/btn_trans_base10.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ripple xmlns:android="http://schemas.android.com/apk/res/android"
+    android:color="@color/base10">
+    <item android:id="@android:id/mask">
+        <shape android:shape="rectangle">
+            <corners android:radius="2dp" />
+            <solid android:color="#fff" />
+        </shape>
+    </item>
+</ripple>
