+++ app/src/main/res/drawable/chip_background.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/base10" />
+    <corners android:radius="16dp" />
+
+</shape>
