+++ app/src/main/res/values/colors.xml
@@ -18,4 +18,9 @@
     <color name="list_choice_normal_bg_light">#ffff</color>
     <color name="list_choice_pressed_bg_light">#fccc</color>
 
+    <!-- Chips -->
+    <color name="paper">#fafafa</color>
+    <color name="base10">#eaeaea</color>
+    <color name="base30">#c1c1c1</color>
+
 </resources>
